本地端数据丢失TAT，尝试恢复进度

+++ src/zzy/model/Author.java
@@ -2,42 +2,51 @@ package zzy.model;
 
 import java.util.*;
 
-public class Author {
-
-    private int AuthorID;
-    private String Name;
-    private int Age;
-    private String Country;
-
-    public void setAuthorID(int AuthorID) {
-        this.AuthorID = AuthorID;
-    }
-
-    public long getAuthorID() {
-        return AuthorID;
-    }
-
-    public void setName(String Name) {
-        this.Name = Name;
-    }
-
-    public String getName() {
-        return Name;
-    }
-
-    public void setAge(int Age) {
-        this.Age = Age;
-    }
-
-    public int getAge() {
-        return Age;
-    }
-
-    public void setCountry(String Country) {
-        this.Country = Country;
-    }
-
-    public String getCountry() {
-        return Country;
-    }
+public class Author
+{
+	
+	private int AuthorID;
+	private String Name;
+	private int Age;
+	private String Country;
+	
+	public void setAuthorID(int AuthorID)
+	{
+		this.AuthorID = AuthorID;
+	}
+	
+	public int getAuthorID()
+	{
+		return AuthorID;
+	}
+	
+	public void setName(String Name)
+	{
+		this.Name = Name;
+	}
+	
+	public String getName()
+	{
+		return Name;
+	}
+	
+	public void setAge(int Age)
+	{
+		this.Age = Age;
+	}
+	
+	public int getAge()
+	{
+		return Age;
+	}
+	
+	public void setCountry(String Country)
+	{
+		this.Country = Country;
+	}
+	
+	public String getCountry()
+	{
+		return Country;
+	}
 }
